# apps/studio/src/integrationTest/resources/openapi/studio.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseIssuedProjectCredentialResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/IssuedProjectCredentialResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorBody"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListProjectCredentialResponse" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/ProjectCredentialResponse"
            },
            "type" : "array"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorBody"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseReleaseResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ReleaseResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorBody"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseRevisionResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/RevisionResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorBody"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ApiResponseSubmissionResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/SubmissionResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorBody"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseVoid" : {
        "properties" : {
          "data" : { },
          "error" : {
            "$ref" : "#/components/schemas/ErrorBody"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "CreatePricingRevisionRequest" : {
        "properties" : {
          "price" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "CreateProjectCredentialRequest" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
...
      "CreateRatingRevisionRequest" : {
        "properties" : {
          "questionnaire" : {
            "additionalProperties" : { },
            "type" : "object"
          }
        },
        "required" : [ "questionnaire" ],
        "type" : "object"
      },
      "CreateReleaseRequest" : {
        "properties" : {
          "publishAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CreateStorePageRevisionRequest" : {
        "properties" : {
          "minimumRequirements" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "platform" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "shortDescription" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "minimumRequirements", "platform", "shortDescription", "title" ],
        "type" : "object"
      },
      "CreateSubmissionRequest" : {
        "properties" : {
          "buildId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "metadataRevisionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pricingRevisionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratingRevisionId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "buildId", "metadataRevisionId", "pricingRevisionId", "ratingRevisionId" ],
        "type" : "object"
      },
...
      "ErrorBody" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "IssuedProjectCredentialResponse" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "gameId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "tokenPrefix" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ProjectCredentialResponse" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "gameId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lastUsedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "tokenPrefix" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ReleaseResponse" : {
        "properties" : {
          "buildId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "channel" : {
            "enum" : [ "LIVE" ],
            "type" : "string"
          },
          "previousReleaseId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "publishAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "releaseId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "SCHEDULED", "PUBLISHED", "CANCELLED" ],
            "type" : "string"
          },
          "submissionId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RevisionResponse" : {
        "properties" : {
          "resolvedPath" : {
            "type" : "string"
          },
          "revisionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "revisionNo" : {
            "format" : "int32",
            "type" : "integer"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SubmissionResponse" : {
        "properties" : {
          "buildId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "gameId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "metadataRevisionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pricingRevisionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "ratingRevisionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "sequenceNo" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "SUBMITTED", "CHANGES_REQUESTED", "READY_FOR_RELEASE", "RELEASED", "CANCELLED" ],
            "type" : "string"
          },
          "submissionId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "모든 응답은 공통 봉투 `{success, data, error}` 로 감싼다.\n실패면 `success=false` 이고 `error.code` 에 `ErrorCode` 이름이 들어간다\n(`PRICE_MISMATCH`, `PAYMENT_AMOUNT_MISMATCH` 등). HTTP 상태는 그 코드가 정한다.\n\n`data` 가 없는 성공 응답(`ApiResponse.ok()`)에서는 `data` 필드 자체가 빠진다 —\n봉투에 `@JsonInclude(NON_NULL)` 이 걸려 있다.",
    "title" : "STOVE studio API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/studio/projects/releases/{releaseId}/cancel" : {
      "post" : {
        "operationId" : "cancel",
        "parameters" : [ {
          "in" : "path",
          "name" : "releaseId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    },
    "/api/v1/studio/projects/releases/{releaseId}/rollback" : {
      "post" : {
        "operationId" : "rollback",
        "parameters" : [ {
          "in" : "path",
          "name" : "releaseId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseReleaseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    },
    "/api/v1/studio/projects/submissions/{submissionId}" : {
      "get" : {
        "operationId" : "submission",
        "parameters" : [ {
          "in" : "path",
          "name" : "submissionId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseSubmissionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    },
    "/api/v1/studio/projects/submissions/{submissionId}/releases" : {
      "post" : {
        "operationId" : "release",
        "parameters" : [ {
          "in" : "path",
          "name" : "submissionId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateReleaseRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseReleaseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    },
    "/api/v1/studio/projects/{gameId}/credentials" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "path",
          "name" : "gameId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListProjectCredentialResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "project-credential-controller" ]
      },
      "post" : {
        "operationId" : "issue",
        "parameters" : [ {
          "in" : "path",
          "name" : "gameId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProjectCredentialRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseIssuedProjectCredentialResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "project-credential-controller" ]
      }
    },
    "/api/v1/studio/projects/{gameId}/credentials/{credentialId}" : {
      "delete" : {
        "operationId" : "revoke",
        "parameters" : [ {
          "in" : "path",
          "name" : "gameId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "path",
          "name" : "credentialId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "project-credential-controller" ]
      }
    },
    "/api/v1/studio/projects/{gameId}/pricing-revisions" : {
      "post" : {
        "operationId" : "pricing",
        "parameters" : [ {
          "in" : "path",
          "name" : "gameId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePricingRevisionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseRevisionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    },
    "/api/v1/studio/projects/{gameId}/rating-revisions" : {
      "post" : {
        "operationId" : "rating",
        "parameters" : [ {
          "in" : "path",
          "name" : "gameId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRatingRevisionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseRevisionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    },
    "/api/v1/studio/projects/{gameId}/store-page-revisions" : {
      "post" : {
        "operationId" : "storePage",
        "parameters" : [ {
          "in" : "path",
          "name" : "gameId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateStorePageRevisionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseRevisionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    },
    "/api/v1/studio/projects/{gameId}/submissions" : {
      "post" : {
        "operationId" : "submit",
        "parameters" : [ {
          "in" : "path",
          "name" : "gameId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSubmissionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseSubmissionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "publishing-controller" ]
      }
    }
  }
}