# apps/studio/src/main/java/com/stove/studio/api/controller/dto/CreateSubmissionRequest.java
package com.stove.studio.api.controller.dto;

import jakarta.validation.constraints.NotNull;

public record CreateSubmissionRequest(
        @NotNull Long metadataRevisionId,
        @NotNull Long pricingRevisionId,
        @NotNull Long ratingRevisionId,
        @NotNull Long buildId
) {
}